Prism Bridge Work + Updates on txt file.
I didnt change too much. Some changes on the txt file so we knew where
we were.
On the world map I added a message stating that leucre was worried about
going onto the bridge but didnt know why.
Also started to create the event on the bridge, but not much.
Added a ton of character generator additions, but these changes are made
on the rpgmaker files not the game files. If you want them let me know
and Ill send them too you.