Added Map, 2 PLugins, Texture and Tileset
Added the Mt. Kloque North Cave that will connect us to the ruins. It's
not necessarily finished, but I feel really limited in our dungeon
tileset and want to expand to add more variety to caves.

Added a new texture for a bird. Haven't used it yet, but it looks nice
and I could see us adding it as just a scenery event to make the world
feel more alive.

Added the bridge to Mt. Kloque connecting the two mountains together and
customized an "Outside_D" tileset to accomodate the edges of the
mountain where the bridge touches. The only purpose of the tileset is to
take into account the 2 specific tiles I needed, will likely add
additional changes to that tileset to justify using an entire tileset
for just 2 tiles.

Now, the juicy stuff!!!!!!

Added two plugins, YEP_ButtonCommandEvent and YEP_KeyboardConfig. These
2 plugins allow us to create custom keybinds, AND bind common events to
key presses.

I currently have the L key toggling our Lantern and the T key brings up
a little menu that lets me toggle the lighting, I'm using that as a sort
of "Master Command" common event that will allow us to on the fly change
things that the in-game switch/variable debug tool won't let us change.

I probably won't be around when you see this, so if during playtesting
none of the hotkeys are working for you, press Escape > Settings >
Keyboard Config and then "Default Keyboard Layout". That should enable
all of the custom hotkeys I've created.

It's super cool! We'll talk more about it tonight if you're available/up
for it!